@@ -11,34 +11,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.1",
-    "@angular/cdk": "^13.1.1",
-    "@angular/common": "~13.1.1",
-    "@angular/compiler": "~13.1.1",
-    "@angular/core": "~13.1.1",
-    "@angular/forms": "~13.1.1",
-    "@angular/material": "^13.1.1",
-    "@angular/platform-browser": "~13.1.1",
-    "@angular/platform-browser-dynamic": "~13.1.1",
-    "@angular/router": "~13.1.1",
-    "@ngrx/effects": "^13.0.2",
-    "@ngrx/entity": "^13.0.2",
-    "@ngrx/store": "^13.0.2",
-    "angular-split": "^13.1.0",
+    "@angular/animations": "^16.2.7",
+    "@angular/cdk": "^16.2.6",
+    "@angular/common": "^16.2.7",
+    "@angular/compiler": "^16.2.7",
+    "@angular/core": "^16.2.7",
+    "@angular/forms": "^16.2.7",
+    "@angular/material": "^16.2.6",
+    "@angular/platform-browser": "^16.2.7",
+    "@angular/platform-browser-dynamic": "^16.2.7",
+    "@angular/router": "^16.2.7",
+    "@ngrx/effects": "^16.2.0",
+    "@ngrx/entity": "^16.2.0",
+    "@ngrx/store": "^16.2.0",
+    "angular-split": "^16.2.1",
     "lodash-es": "^4.17.21",
-    "ol": "^6.9.0",
-    "ol-ext": "^3.2.18",
-    "ol-geocoder": "^4.1.2",
-    "proj4": "^2.7.5",
-    "rxjs": "~7.4.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "ol": "^8.1.0",
+    "ol-ext": "^4.0.11",
+    "ol-geocoder": "^4.3.1",
+    "proj4": "^2.9.0",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.6.2",
+    "zone.js": "~0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.2",
-    "@angular/cli": "~13.1.2",
-    "@angular/compiler-cli": "~13.1.1",
-    "@angular/language-service": "~13.1.1",
+    "@angular-devkit/build-angular": "^16.2.4",
+    "@angular/cli": "^16.2.4",
+    "@angular/compiler-cli": "^16.2.7",
+    "@angular/language-service": "^16.2.7",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash-es": "^4.17.3",
@@ -57,6 +57,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.5.4"
+    "typescript": "~5.1.6"
   }
 }
