@@ -1,11 +1,9 @@
 import {Component, HostBinding, OnInit} from '@angular/core';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormBuilder, UntypedFormGroup, Validators} from '@angular/forms';
 import {Router} from '@angular/router';
-import {MatSnackBar} from '@angular/material/snack-bar';
+import {MatLegacySnackBar as MatSnackBar} from '@angular/material/legacy-snack-bar';
 import {EMAIL_REGEX} from '../../_helpers/regex';
 import {ApiService} from '../../_services/api.service';
-import {catchError} from 'rxjs/operators';
-import {of} from 'rxjs';
 
 @Component({
   selector: 'gs2-forget',
@@ -17,13 +15,13 @@ export class ForgetComponent implements OnInit {
   @HostBinding('class') class = 'main-container';
 
   private readonly successMessage = 'Le mot de passe a été envoyé à l\'adresse : ';
-  form: FormGroup;
+  form: UntypedFormGroup;
 
   get email() {
     return this.form.get('email');
   }
 
-  constructor(private fb: FormBuilder,
+  constructor(private fb: UntypedFormBuilder,
               private router: Router,
               public snackBar: MatSnackBar,
               private apiService: ApiService
