@@ -2,15 +2,15 @@ import {Component, OnInit, ElementRef, ViewChild} from '@angular/core';
 import {IProduct} from 'src/app/_models/IProduct';
 import {ApiService} from 'src/app/_services/api.service';
 import {ConfigService} from 'src/app/_services/config.service';
-import {MatDialog} from '@angular/material/dialog';
+import {MatLegacyDialog as MatDialog} from '@angular/material/legacy-dialog';
 import {DialogMetadataComponent} from './dialog-metadata/dialog-metadata.component';
-import {FormControl} from '@angular/forms';
+import {UntypedFormControl} from '@angular/forms';
 import {BehaviorSubject, merge, Observable} from 'rxjs';
 import {debounceTime, map, mergeMap, scan, switchMap, tap, throttleTime} from 'rxjs/operators';
 import {CdkVirtualScrollViewport} from '@angular/cdk/scrolling';
 import {AppState, selectOrder} from '../../_store';
 import {Store} from '@ngrx/store';
-import {MatSnackBar} from '@angular/material/snack-bar';
+import {MatLegacySnackBar as MatSnackBar} from '@angular/material/legacy-snack-bar';
 import {GeoshopUtils} from '../../_helpers/GeoshopUtils';
 import {IOrder} from '../../_models/IOrder';
 import {updateOrder} from '../../_store/cart/cart.action';
@@ -33,7 +33,7 @@ export class CatalogComponent implements OnInit {
   isSearchLoading = false;
 
   // Filtering
-  catalogInputControl = new FormControl('');
+  catalogInputControl = new UntypedFormControl('');
 
   mediaUrl: String | undefined;
   order: IOrder;
