@@ -4,13 +4,12 @@ import {Store} from '@ngrx/store';
 import * as fromCart from '../../_store/cart/cart.action';
 import {IProduct} from '../../_models/IProduct';
 import {DialogMetadataComponent} from '../../welcome/catalog/dialog-metadata/dialog-metadata.component';
-import {MatDialog, MatDialogRef} from '@angular/material/dialog';
+import {MatLegacyDialog as MatDialog, MatLegacyDialogRef as MatDialogRef} from '@angular/material/legacy-dialog';
 import {ApiService} from '../../_services/api.service';
-import {MatSnackBar} from '@angular/material/snack-bar';
+import {MatLegacySnackBar as MatSnackBar} from '@angular/material/legacy-snack-bar';
 import {ConfirmDialogComponent} from '../confirm-dialog/confirm-dialog.component';
 import {Router} from '@angular/router';
 import {ApiOrderService} from '../../_services/api-order.service';
-import {takeUntil} from 'rxjs/operators';
 import {Subject} from 'rxjs';
 import {StoreService} from '../../_services/store.service';
 import {MapService} from '../../_services/map.service';
