@@ -1,12 +1,12 @@
 import {HttpInterceptor, HttpRequest, HttpHandler, HttpEvent, HttpErrorResponse} from '@angular/common/http';
-import {Observable, of, throwError} from 'rxjs';
+import {Observable, throwError} from 'rxjs';
 import {Injectable} from '@angular/core';
 import {AppState} from '../_store';
 import {Store} from '@ngrx/store';
 import * as fromAuth from '../_store/auth/auth.action';
 import {catchError} from 'rxjs/operators';
 import {Router} from '@angular/router';
-import {MatSnackBar} from '@angular/material/snack-bar';
+import {MatLegacySnackBar as MatSnackBar} from '@angular/material/legacy-snack-bar';
 
 @Injectable()
 export class ErrorInterceptor implements HttpInterceptor {
