@@ -7,7 +7,7 @@ import { IApiResponse } from '../_models/IApi';
 import { catchError, flatMap, map } from 'rxjs/operators';
 import { Contact, IContact } from '../_models/IContact';
 import { GeoshopUtils } from '../_helpers/GeoshopUtils';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatLegacySnackBar as MatSnackBar } from '@angular/material/legacy-snack-bar';
 
 
 @Injectable({
@@ -274,7 +274,7 @@ export class ApiOrderService {
   getContact(contactId: number | string) {
     this._getApiUrl();
 
-    if (contactId < 0 || typeof contactId !== 'number') {
+    if (typeof contactId !== 'number' || contactId < 0) {
       return of(null);
     }
     const url = new URL(`${this.apiUrl}/contact/${contactId}/`);
