@@ -7,8 +7,8 @@ import {catchError, exhaustMap, map, take, tap} from 'rxjs/operators';
 import {ApiService} from '../../_services/api.service';
 import {of} from 'rxjs';
 import {Router} from '@angular/router';
-import {MatSnackBar, MatSnackBarRef} from '@angular/material/snack-bar';
-import {MatDialog, MatDialogRef} from '@angular/material/dialog';
+import {MatLegacySnackBar as MatSnackBar, MatLegacySnackBarRef as MatSnackBarRef} from '@angular/material/legacy-snack-bar';
+import {MatLegacyDialog as MatDialog, MatLegacyDialogRef as MatDialogRef} from '@angular/material/legacy-dialog';
 import {ConfirmDialogComponent} from '../../_components/confirm-dialog/confirm-dialog.component';
 import {ApiOrderService} from '../../_services/api-order.service';
 import {IOrder} from '../../_models/IOrder';
