@@ -1,6 +1,6 @@
 import {Component, OnInit, Inject} from '@angular/core';
 import {ConfigService} from 'src/app/_services/config.service';
-import {MatDialogRef, MAT_DIALOG_DATA, MatDialog} from '@angular/material/dialog';
+import {MatLegacyDialogRef as MatDialogRef, MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA, MatLegacyDialog as MatDialog} from '@angular/material/legacy-dialog';
 import {IMetadata} from 'src/app/_models/IMetadata';
 
 @Component({
